+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix unterminated string buffer in dc1 error handling routine (#28)
+- dupdate change to support wildcards broke Kickstart 1.3 compatibility
 
 ### Removed